+++ htdocs/societe/class/societe.class.php
@@ -83,7 +83,8 @@ class Societe extends CommonObject
 	);
 
 	/**
-	 * @var array	List of child tables. To know object to delete on cascade.
+	 * @var array    List of child tables. To know object to delete on cascade.
+	 *               if name like with @ClassNAme:FilePathClass:MethodDetele' it will call method to delete object rather tahn SQL delete
 	 */
 	protected $childtablesoncascade = array(
 		"societe_prices",
@@ -92,7 +93,7 @@ class Societe extends CommonObject
 		"product_fournisseur_price",
 		"product_customer_price_log",
 		"product_customer_price",
-		"socpeople",
+		"@Contact:/contact/class/contact.class.php:deleteBySoc",
 		"adherent",
 		"societe_account",
 		"societe_rib",
@@ -104,13 +105,6 @@ class Societe extends CommonObject
 		"notify_def",
 		"actioncomm",
 	);
-
-	/**
-	 * Build as [parentChildTable]=>[childTable]=>array(parentChildTable.PrimaryKey,childTable.ForeignKey,WhereClause)
-	 *  Define into Contructor because of MAIN_DB_PREFIX cannot be use here
-	 * @var array	List of Parent's child's child tables. To know object to delete on cascade.
-	 */
-	protected $parentchildchildtablesoncascade = array();
 
 	/**
 	 * @var string String with name of icon for myobject. Must be the part after the 'object_' into object_myobject.png
@@ -666,22 +660,6 @@ class Societe extends CommonObject
 		$this->forme_juridique_code = 0;
 		$this->tva_assuj = 1;
 		$this->status = 1;
-
-		$this->parentchildchildtablesoncascade = array(
-			'socpeople'=>
-				array(
-					MAIN_DB_PREFIX.'socpeople_extrafields' =>
-				      array(MAIN_DB_PREFIX.'socpeople.rowid',MAIN_DB_PREFIX.'socpeople_extrafields.fk_object',''),
-				    MAIN_DB_PREFIX.'element_contact' =>
-				      array(MAIN_DB_PREFIX.'socpeople.rowid',
-				            MAIN_DB_PREFIX.'element_contact.fk_socpeople',
-				            MAIN_DB_PREFIX.'element_contact.fk_c_type_contact IN (SELECT ct.rowid FROM '.MAIN_DB_PREFIX.'c_type_contact as ct WHERE ct.source=\'external\')'),
-					MAIN_DB_PREFIX.'societe_contacts' =>
-						array(MAIN_DB_PREFIX.'socpeople.rowid',
-						      MAIN_DB_PREFIX.'societe_contacts.fk_socpeople',
-						      MAIN_DB_PREFIX.'societe_contacts.fk_c_type_contact IN (SELECT ct.rowid FROM '.MAIN_DB_PREFIX.'c_type_contact as ct WHERE ct.source=\'external\')'),
-					)
-		);
 	}
 
 
@@ -1702,30 +1680,31 @@ class Societe extends CommonObject
 
 			foreach ($this->childtablesoncascade as $tabletodelete)
 			{
-				if (!$error && array_key_exists($tabletodelete, $this->parentchildchildtablesoncascade))
-				{
-					if (count($this->parentchildchildtablesoncascade[$tabletodelete])>0){
-						foreach($this->parentchildchildtablesoncascade[$tabletodelete] as $childtabletodelete=>$dataToDelette) {
-							$sql = "DELETE FROM ". $childtabletodelete;
-							$sql .= " WHERE ".$dataToDelette[1]." IN (SELECT ".$dataToDelette[0]." FROM ".MAIN_DB_PREFIX.$tabletodelete." WHERE fk_soc = " . $id.")";
-							if (!empty($dataToDelette[3])) {
-								$sql .= " AND ".$dataToDelette[3];
-							}
-							if (!$this->db->query($sql)) {
-								$error++;
-								$this->errors[] = $this->db->lasterror();
-							}
-						}
-					}
-				}
 				if (!$error)
 				{
-					$sql = "DELETE FROM ".MAIN_DB_PREFIX.$tabletodelete;
-					$sql .= " WHERE fk_soc = ".$id;
-					if (!$this->db->query($sql))
-					{
-						$error++;
-						$this->errors[] = $this->db->lasterror();
+					$delete_from_object=explode(':', $tabletodelete);
+					if (count($delete_from_object)>1) {
+						$class_name=str_replace('@', '', $delete_from_object[0]);
+						$filepath=$delete_from_object[1];
+						$delete_method=$delete_from_object[2];
+						if (dol_include_once($filepath)) {
+							$child_object = new $class_name($this->db);
+							$result= $child_object->{$delete_method}($id);
+							if ($result<0) {
+								$error++;
+								$this->errors[] = $child_object->error;
+							}
+						} else {
+							$error++;
+							$this->errors[] = 'Cannot find child class file ' .$filepath;
+						}
+					} else {
+						$sql = "DELETE FROM " . MAIN_DB_PREFIX . $tabletodelete;
+						$sql .= " WHERE fk_soc = " . $id;
+						if (!$this->db->query($sql)) {
+							$error++;
+							$this->errors[] = $this->db->lasterror();
+						}
 					}
 				}
 			}
